pyproject.toml
[tool.poetry]
name = "datapane_challenge"
version = "0.1.0"
description = ""
authors = ["Sergi Chalauri <...>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio-tools = "~=1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
